requirements.txt
folium==0.20.0
langchain==1.2.15
openai==2.31.0
pydantic==2.12.5
python-dotenv==1.2.2
Requests==2.33.1
streamlit==1.56.0
streamlit_folium==0.27.1
